+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ﻿## [Unreleased] ##
 ### Fixed ###
 
+  * [retail] Missed some textures on the spec frame
   * [retail] Display bug with quest detail
   * [retail] World quest header wasn't skinned
   * [retail] Display bug with the calendar