+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ﻿## [Unreleased] ##
+### Changed ###
+
+  * Tweaked font for chat bubble names
+
 ### Fixed ###
 
+  * Chat bubble didn't work
   * [retail] Missed some textures on the spec frame
   * [retail] Display bug with quest detail
   * [retail] World quest header wasn't skinned