+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
   * Error when loot skin is disabled
   * Chat bubble didn't work
+  * [retail] PvP category buttons had off borders
   * [retail] Sell item icon had the wrong texture
   * [retail] Volume sliders were white
   * [retail] Missed some textures on the spec frame