+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-﻿## [Unreleased] ##
+﻿## [9.0.1.5] - 2020-10-25 ##
 ### Changed ###
 
   * Tweaked font for chat bubble names
@@ -224,6 +224,7 @@
   * Error when opening azerite respec UI
 
 [Unreleased]: https://github.com/Haleth/Aurora/compare/master...develop
+[9.0.1.5]: https://github.com/Haleth/Aurora/compare/9.0.1.4...9.0.1.5
 [9.0.1.4]: https://github.com/Haleth/Aurora/compare/9.0.1.3...9.0.1.4
 [9.0.1.3]: https://github.com/Haleth/Aurora/compare/9.0.1.2...9.0.1.3
 [9.0.1.2]: https://github.com/Haleth/Aurora/compare/9.0.1.1...9.0.1.2